Convert callouts to section variants
