--- stories/components/callout/callout.content.json
@@ -1,10 +1,0 @@
-{
-  "image": "https://www.bixal.com/static/img-whatwedo-e68b3d20abd32c896d56b122063f7664.jpg",
-  "prefix": "What we do",
-  "title": "The work we do helps our clients unite stakeholders, optimize resources, and better serve citizens all over the world.",
-  "description": "We take a people-absolutely-first approach to solving complex organizational challenges and gracefully balance cutting-edge technical chops with a deep sense of empathy and understanding. We relentlessly focus on outcomes and weave it all together with a unique agility that permeates across everything we do.",
-  "cta": {
-    "href": "javascript:void(0)",
-    "label": "Explore case studies"
-  }
-}
