+++ stories/components/section/section.scss
@@ -3,6 +3,10 @@
 // @TODO Create mixin.
 .bix-section {
   padding: 100px 0;
+
+  &__footer {
+    margin-top: units(6);
+  }
 }
 
 .bix-section--base {
@@ -39,5 +43,13 @@
     &::placeholder {
       color: rgba(white, 0.6);
     }
+  }
+}
+
+.bix-section--centered {
+  text-align: center;
+
+  .bix-section__title::after {
+    content: none;
   }
 }
